Ignore `withdrawn` when checking translations
Due to a bug in the Publishing API, `withdrawn` content doesn't contain
an `available_translations` links element.

While this is the case and as it doesn't have a massive impact (or any
at the moment because there aren't any migrated formats with translated
withdrawn content) we should ignore them when checking the translations.