Add `TopicContentIdMap`
This class fetches the `content_id` for all `SpecialistSector` objects
and returns them as a hash of `base_path => content_id`. The result is
memoized within a class method so will not pick up any changes until the
process is restarted. This is intentional is it is to be used within the
sync checks and will prevent having to make a publishing api request for
every edition checked. There are < 400 distinct base paths so memory
consumption is not an issue.