Delete `EditionPolicy` with non-existent `policy_content_id`
There are 614 `EditionPolicy` records that have a `policy_content_id`
that does not exist in PublishingAPI. This data migration deletes them.
To preserve the history in the event that the related content comes to
light I've listed all of the objects as an array of hashes that could be
used to reinstate the objects.