Republish `DetailedGuide`
This will republish all `DetailedGuide` documents through the updated
worker. This has been done numerous times on integration and staging but
we now need to run it on production to do the final sync checks before
switching over the rendering app.

Whilst we're not switching over the rendering app yet, this republishing
will appropriately create redirect routes for unpublished items in the
router so some editions will at that point be migrated away from
Whitehall.