Set OperationalFieldPresenter update_type to major
Publishing API requires an `update_type`. In addition, the first
published state *must* be major for the Publishing API to accept it.

For `OperationalField` `update_type` cannot currently be specified in
the Whitehall UI. Thus it should always be major.

In order to satisfy the interface required by the workers the
`OperationalFieldPresenter` ignores the `update_type` keyword argument.