Send logo url for Detailed Guides if defined
One detailed guide has a `logo_url` configured for regulatory reasons.
This should be sent to the publishing api so that government frontend
can render it.
It isn’t possible for Detailed guides to have a main image in any other
way, so we reuse the `image` field in the details hash for this purpose