Use content_id to refer to topics
The SpecalistSector model holds the relationship between an Edition and
a "topic" (formerly specialist sector). It uses the slug of the topic
as a reference. Since these are subject to change the new world
publishing-api uses the content_id to refer to other items.

This commit adds a `topic_content_id` to the table and populates it
with the existing topics. The lookup-list is hardcoded statically to
make the migration resilient against publishing-api timeouts.

There are currently ~250 records that refer to topics that no longer
exist yet (thus proving the usefulness of this change). We'll handle
these in a different commit.