Publish the new finder page for people via rake
We are migrating the rendering app of the people page from `whitehall` to
`finder-frontend`. In doing so, we need a way to publish the new finder to the
publishing API.

This commit adds the new finder for the people page that can be published via a
rake task to the publishing API.

It also adds a test case so that we are sure that the schema is valid.