Send unpublished_at to Publishing API on withdrawal
Whitehall displays the last updated time of a withdrawn edition as the
date it was withdrawn. This wasn’t being passed through to Publishing
API and therefore Government Frontend wasn’t able to display it.

This uses new functionality in Publishing API and passes what Whitehall
considers to be the withdrawn date when unpublishing or republishing.