Only present valid specialist sectors to Publishing API
It is possible for the relationship between a specialist sector and an edition
to remain persisted despite the specialist sector being invalid. This extra
check ensures we only present valid sepcialist sectors to the Publishing API.

This change was identified when publishing a document collection that had
a specialist sector that was missing a `content_id`. The specialist sector
in question appears to have been withdrawn as it is not visible on GOV.UK.

The Publishing API rejected the JSON as the links hash contained the `nil`
`content_id` of the specialist sector.