Add model for storing sync check results
We store the check class name instead of the class of item due to the
polymorphism of the individual records that are being checked. For
example, for all editioned items, the class would be Document, which
doesn’t provide enough information about the type of document being
checked.

Serialising the failures using YAML allows the results to be read back
in as  the original SyncChecker::Failure struct. This isn’t optimal,
but this model is intended to be used by developers only to assist with
migration of documents.