Add SyncCheckWorker
Takes a check class (from SyncChecker::Formats) and an id. Will run the
sync check for that item and record the results in the
sync_check_results table.

The check class can be passed as either a class or string, for
compatibility with running directly in the console or via Sidekiq.

Records the failures as NULL if the check was successful for easier
filtering in the database.