Fix date comparison in unpublished sync check
Date formats are not consistent across our applications.
Sync checks on unpublished items are failing because withdrawn dates
in Content Store are returned as UTC while Whitehall records them
as a DateTime with an offset (this is in the `created_at` field).

Converting both dates to UTC means we are comparing like for like
and also takes care of adding or subtracting our offset for free.

The sync check tests now also reflect this change and test
against UTC representations of dates rather than rfc3339 as before