Exclude draft topics from edition sync checks
Draft topics can be associated with an edition. This association should be
presented to the Publishing API so it can maintain that assocation and
perform any operations on linked content items should the draft topic
be modified in the future.

However, we are unable to assert that association exists through a
sync check as draft topics are excluded from the content item we
assert against. The content items represent to the public state
of an edition on GOV.UK and not the state of the edition private
to the publishing platform.