Fix related documents for two document collections
Two of the document collections we are migrating are linking to
documents which are in an odd state in Whitehall; their latest edition
is published, but the one before that is draft:

```
> d = Document.find(168857)
> d.editions.map(&:state)
["superseded", "draft", "published"]
```

This causes `PublishingApiDocumentRepublishingWorker` to pick up the
draft as a `pre_publication_edition`, and since it has an
unpublishing, that will lead to a redirect being published to the
content-store. This will fail the sync check.

This data migration finds the culprits and sets them to superseded,
while deleting their unpublishings.

Investigated by @gpeng and @mgrassotti, I'm just writing the data
migration.