Redraft and redirect unpublished-fishing-focus
This collection has been renamed as a means to freeing up the base path
for a new piece of content, however this doesn't get reflected in the
Publishing API so the renamed item needs a new content_id and
redrafting then unpublished as a redirect to the newer content item.
It's that simple.