Send topical event start and end dates to rummager
These fields were added to rummager here:
https://github.com/alphagov/rummager/pull/715

This allows us to build a finder for topical events that knows about
active/inactive topical events.

In whitehall, an active topical event is one with an end date that is in the
future. All others are considered inactive and are not listed at the moment.
We order topical events by the start date.

We display topical events and policy areas at /government/topics.
Replacing with finders is the simplest way to generate this page via the
new publishing platform and means we can delete the action from whitehall.

https://trello.com/c/oVfSAO9W/311-migrate-policy-areas-index-to-a-finder