Refactor some of the rummager indexing code
The aim is to eventually get rid of the `Searchable` mixin and create presenter
classes that can be used with gds-api-adapters instead of rummageable.

The search indexing code has two ways of doing the same thing and both
rely on a deprecated gem (rummageable). Normally we index content via
the sidekiq workers, but some rake tasks skip this and run the indexing
synchronously.

The indexing code is also very intertwined with the model code at
the moment - it would be easier to work with if we could separate
this logic out, as we have done for publishing api.

I think this is worth addressing now because the existing code is so
unintutive. `#search_index` is not a search index - it's a collection
of payloads for rummager.

At some point in the future I expect us to get to be able to index all
content directly from the publishing platform and this integration can
go away - but this is not going to happen soon.

This commit moves rummager functionality out of the "Whitehall" module.
I've moved methods for generating rummager payloads for multiple content
items to a new SearchPresenters namespace. I decided not to define the
names of the indexes here - I've moved that to the
Whitehall::SearchIndex class instead. This entire class should be
removed and replaced with GDS API adapters.

SearchPresenters also contains the logic for deciding which classes go
to which index - which is mostly pointless because DetailedGuides is the
only exception and there is no longer any meaningful distinction between
how the indexes are used. At some point these will be combined.  The
only other place that reference these methods is a `rummager_export.rb`
script.  I suspect this script is not useful anymore either but I've
left it alone for now.

I've also changed some Plek calls to use `rummager` rather than
`search` since the `search` vhost is deprecated: see also
https://github.com/alphagov/gds-api-adapters/commit/ddae94d1656f59f7778ca794a0cce356fc75e7e2

This commit leaves behind all the `#search_index` methods in the models,
that present a single content item to rummager. That should be addressed
in a separate PR.

https://trello.com/c/BvriiajH/328-whitehall-search-indexing-tech-debt