Finder for topical events
This splits the list of topical events from the list of policy areas, which are
currently combined on the /government/topics index page.

We're creating finders for both of these, which adds search/filtering
functionality and removes the need for additional templates and controllers in
whitehall.

The finder can be filtered by status (open or closed) and the date (`open_date`).

https://trello.com/b/GBkshanv/taxonomy-tea://trello.com/c/oVfSAO9W/311-migrate-policy-areas-index-to-a-finder