PublishingApiHtmlAttachmentWorker handles `unpublished`
Ok... Here goes... When we unpublish an HtmlAttachment as a result of a
republishing action we need to send the `allow_draft: true` argument to
the 'next' worker and also set `update_type: "republish"`.
`allow_draft` is required as there are a number (~300) of corresponding
`ContentItem` records in publishing-api that are in draft state and can't be
unpublished otherwise.

This commit makes `republish` also handle `unpublishing`. This action will
be called from the `PublishingApiDocumentRepublishingWorker` when the parent
attachable `Edition` is republished.