Remove hackable URL test
This commit removes the “hackable URL” test, which asserts that various whitehall-served URLs are “hackable” all the way up to `/government` and serve useful index pages. With the migration of the whitehall finders to finder-frontend, there are very few pages left for this test to look at.

The only remaining valid un-migrated paths are:

* /government/ministers
* /government/fields-of-operation
* /government/world
* /courts-tribunals