Sort and group world locations to respect accented characters
In 3c0eb7e2a9f4ef791a9403c6d3f2fda8bad0590b we changed the sort order of
world locations on the public index page to ignore "The" prefixes, so that
places like "The Occupied Palestinian Territories" appear under "O" not
"T".  Unfortunately this meant sorting in ruby and this didn't respect
accented characters, meaning places like "São Tomé and Principe" appeared
at the end of "S" rather than inbetween "San Marino" and "Saudi Arabia".

The order before we account for "The" prefixes is correct for accented
characters, because it comes from the DB which has the correct collation
settings.  The version of the name without the prefix does not live in the
DB so we have to do the sort in memory, and this breaks the ordering for
accented characters.

The solution, much like that used in Frontend (https://github.com/alphagov/frontend/pull/1053)
is to use `ActiveSupport::Inflector.transliterate` to convert accented
characters to their de-accented variants and sort using that version of
the name.  This makes the grouping marginally more complex as we also
use the first char of the prefix-less name to group by, and we don't want
to create a group for the accented characters, and we don't want to
transliterate the name twice either.