Discard deleted HtmlAttachment drafts on update
When a draft `Edition` is saved any attachments that have been deleted
need to have their drafts discarded in the publishing API or else they
will be left as they are ignored when the draft is published.

This is handled in the `delete` command but that is not called in the
edition lifecycle.