Remove services and information rendering code
This commit removes “services and information” pages rendering code from whitehall, now that the pages are rendered by collections (as of https://github.com/alphagov/collections/pull/197). Code still exists in whitehall to publish the content item for these pages either via a rake task or when an organisation is updated.