Default to the user organisation in filters
In the organisation features, the default organisation wasn't set,
meaning we would be showing editions for all organisations.

That query is quite slow and is causing issues in mysql in production.

In order to mitigate that, we are filtering the editions by the users'
organisation by default, which makes the database queries faster on the
first page hit.

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/1472963

Errbit that shows the slow queries:
https://errbit.publishing.service.gov.uk/apps/53020d6c0da11585f10000e7/problems/583d970965786364a5f10400