Remove references to topics_path
Whitehall no longer serves an index page for topics and topical events.  This
conflicts with a lot of code (mostly tests) that thinks these should exist.

`ApplicationHelper` contains a method `current_main_navigation_path` that
atempts to work out a home for various pages, based on the controller handling
them and the query params.

The only placed it's used is when comparing to the various top level
nav pages in whitehall (to tell whether to highlight a button)

So the only useful return values are the ones corresponding to these:
- Departments
- Worldwide
- How government works
- Get involved
- Policies
- Publications
- Consultations
- Statistics Announcements

It's useless for policy area/topical events because `/topics` is not a top
level navigation page, and it never will be.

Also, we now have a `parent` link in the content store that will tell
you where a page belongs, so this kind of thing shouldn't be in a
rendering app at all.