Remove re-editioned edition
This piece of content was archived and then re-editioned around the
time of the 2015 election. The two editions differ in that the live
withdrawn edition mentions the English Heritage organisation and the
draft we're going to remove mentions the Historic England org.
Historic England replaced English Heritage around March 2015 but the
draft content updates were never published, we're unsure why.
To allow sync checks to proceed we need to remove this draft as it's
a workflow anomally.