Update content_id to match Publishing API
A number of `HtmlAttachment` objects have mis-matching `content_id` when
compared to the content at the same base path in Publishing API.

This was probably caused by us not deleting draft items when the
attachments were deleted so when a new attachment is created with the
same base path and gets a new `content_id` in Whitehall a conflict
occurs.

This commit also updates a few `Publication` records that suffer the
same conflict. Again, probably caused by issues when we first started
sending them.