Provide translation for blog_post document type
Without a translation we humanize and titleize the document type so we
get "Blog Post" but this would contravene the GOV.UK style guide so we
provide this translation to give us "Blog post" instead.  We only have
the English translation so the others will fall back to this, although
we do add the empty translations to avoid causing the i18n_key_test to
fail.