Add "active" state to virtual tour links
The intention is to make this list look more like the list of previous
prime ministers by having an "off" state for the current tour.  The prime
ministers page does this serverside by having that list entry not be a
link.  We have to do it client-side by adding a class and having that
change the styling to no longer look like a link.