Send the organisation type to rummager
In https://github.com/alphagov/whitehall/pull/2875 I attempted to send
the organisation type to rummager, so that we can make a finder for
organisations. I forgot to add the new field to the `Searchable`
module, so it didn't actually work. The lack of test changes should
have tipped me off.