Use asset pipeline path helpers for virtual tour files
The SWF and XML files are in the asset pipeline already, but we refer to
them directly.  This means the tour only works on dev (where assets don't
get compiled into hash urls) or on production (where the unadorned files
have been uploaded to the whitehall frontend machines directly).

If we use asset_path helpers instead the tour should work in all
environments.

We also have to update the xml files used by the tour to use asset helpers
too as these contain urls to skin files and images used to present the UI
of the pano.