Fix All Organisations filter
Trello:
https://trello.com/c/NE2xfVo8/539-bad-gateway-error-for-featured-documen
ts-in-whitehall-admin-1-day

Selecting All Organizations was leading to a server timeout and a
subsequent Bad Gateway error. This is probably due to the query which
is taking too long.

As a temporary solution we are inviting the user to insert a title or a
slug if “All organisations” is selected, to reduce the target scope.