Refactor `HasContentId` module
Moved the creation of the `uuid` into a method that can be overridden in
classes that include the module. HtmlAttachments need to be able to set
the content id to that of a previously deleted instance if the titles
match so this will allow the method to be overriden in that class.