HtmlAttachments retain the same content_id
When a published HtmlAttachment is deleted the base path is
redirected to the parent document's base path. If another attachment is
created that has the same base path it is currently assigned a new
`content_id`. This causes it to be rejected by PublishingApi as there is
already a redirect content item occupying that base path with a
different `content_id`.

This commit checks for a previously published HtmlAttachment at the same
base path when creating a new one and persists that previous
`content_id` if one is found.