Add script for generating report for Content Operating Model team
The report is to show the organisation, url, format, status, date of
first publication, and date of the last change to the status for all
the editioned documents in the system.

We expect that this is a one-off, but it might be useful for the code
that generates it to persist as this kind of report could be needed
again, or could be used as a jumping off point for future reports.

The trello card describing the report is: https://trello.com/c/e3nDUdE2/88-report-for-content-operating-model