Allow repeated unpublishing transitions
Allows the edition state `withdrawn` to transition to `withdrawn` and
`draft` (the state that the other types of unpublishing end up in) to
transition to `draft`.

To support updating unpublishing/withdrawn messages we need to call the
unpublish endpoint on the publishing API multiple times. In order that
other things (such as also unpublishing HtmlAttachments) occur at the
same time we need to do this via the EditionServiceCoordinator rather
than just calling the publishing api directly. This requires that we
allow the unpublishing transitions to occur more than once
e.g. `draft` -> `draft` and `withdrawn` -> `withdrawn`