Env var to disable config.assets.debug in dev
This setting can slow down page render times considerably, because there
are over 100 different javascript files that get requested individually.
In some cases the page can take around 20 seconds just to finish
rendering.  This really slows down development if you're not working on
the frontend.

> When debug mode is off, Sprockets concatenates and runs the necessary
> preprocessors on all files.
http://guides.rubyonrails.org/asset_pipeline.html#turning-debugging-off

Debug mode is on by default because with it off, compiled assets are
cached, which interferes with auto reloading.