Merge rummageable into whitehall
whitehall is the only remaining app that uses the rummageable gem. Due to the complexity of re-writing the search code to use gds-api-adapters and the current nature of flux in the codebase, this commit merges the rummageable gem into the whitehall codebase. This allows the gem to be fully deprecated while also allowing whitehall to continue to utilise the code in a self-contained manner.