Fix state of collision-at-shalesmoor-tram-stop
The Document with the slug "collision-at-shalesmoor-tram-stop" is
causing errors on production. The Document contains one Edition which
has a state of "withdrawn", however there are no Unpublishings for this
Edition. This causes an error when the `_withdrawn_notice.html.erb`
partial is rendered since it expects an `unpublishing` to be present.

This migration sets the state of the Edition to "published" which will
stop this error.