Fix published Consultation first_published_at
https://github.com/alphagov/whitehall/pull/2959 introduced changes to
fix the `first_public_at` value to be when the Consultation was
published not when it opened. It however failed to backfill the
existing Consultation data so that the correct values are in place so
that the inherited code from Edition behaves as expected.

Consultation previously set `make_public_at(date)` as a NOOP causing
all existing Consultations to have a nil `first_published_at` value
when they were published. This migration backfills the Consultations
to have the value that would have been returned had the code not been
removed.

We have to skip validations as they expect the date to be in the present.