Hierarchy of checkboxes tagging interface
This commit replaces the autocomplete-enabled text input with a bunch of
checkboxes arranged to show the structure of the taxonomy.

Since our guidance is to always tag to the most specific topic possible,
we don't allow any states where both a parent taxon and one of its child
(or grandchild etc) taxons are simultaneously selected. If the server
receives both, the parent is ignored.

On the front end, we show this differently. Whenever the child is
selected, all of its ancestors are shown as selected. This doesn't
affect what is passed to the publishing API. It's just to indicate that
any of those ancestors could show the content on their navigation pages.
In the current navigation prototype, a topic page shows an accordion of
all the content tagged to its children if none of those child taxons
themselves have child taxons.

Trello:
https://trello.com/c/q9imXBPw/441-implement-the-tree-tagging-interface-in-whitehall
and
https://trello.com/c/h7TGge8t/465-change-the-tree-tagging-interface-in-whitehall-to-have-symmetric-checkbox-interactions
Paired with @klssmith @Davidslv @rubenarakelyan