Resolve content id conflicts...
Some workflow to retire content and succeed it with new content
has caused a base_path conflict in the publishing API for a couple
of publications. In both cases one of the publications is redundant
and the publishing API has a record of it, so any attempts to republish
the useful content fails with a conflict. So repurpose the blocking content
id and remove the redundant document in Whitehall so that republishing works.