Add `EXECUTOR_NUMBER` to the test database name
We currently have Whitehall pinned to one agent in CI and this is
causing issues when two or more builds run concurrently. Each time a
build starts it drops the test databases and as these aren't currently
namespaced the already running build then fails.

This commit adds the EXECUTOR_NUMBER environment variable to the DB names with
which should prevent this condition occurring.