Harden whitelisted urls validation for offsite links
We want to let people use the whitelisted domains, but we need to be sure
it's only the whitelisted domains, not things that look like them. So we
make sure the whitelisted domain is either preceded by the start of the
string, or by a "." indicating a subdomain (which we allow).

e.g. www.nhs.uk is ok, but not-the-nhs.uk is not.