Connect type dropdown in UI and model validations for OffsiteLinks
In the UI the dropdown for allowable types was a hand-coded list that
needs to be kept manually in sync with the list of allowable values for
the link_type attribute.  By extracting the types to a separate concern
we can connect the source for the validation and the source for the
values in the dropdown.