Fix sync checker edition base `get_path`
The SyncChecker::Formats::EditionBase creates base paths which it uses
to query the content store. When an edition with a primary_locale not
equal to english is being checked, however, curious base paths are
being created which contain the appended locale twice. For example the
base path returned is
government/world-location-news/172185.fr.fr instead of
government/world-location-news/172185.fr

This is caused by the `get_path` method which uses Whitehall::UrlMaker
to construct the base path and then appends the locale if the locale
being sync checked is not "en".

Unfortunately, UrlMaker automatically appends the locale to a path if
the edition is a `non_english_edition?`. This occurs in editions such
as World Location News Articles where an edition can be designated as
being foreign language only. In the database, this is reflected by the
edition having a `primary_locale` not equal to english.