Fix attachment locale filtering
`Publication`s can be created in multiple languages. Associated
`HtmlAttachment`s can also be created in multiple languages and
Whitehall allows the publisher to choose which language version of a
`Publication` to attach to, either a specific version or 'all
languages'.

This commit updates the `PublishingApi::PublicationPresenter` to support
that. Previously we were treating an attachment with `nil` locale as
being 'default' ie `:en`.