Republish the latest 10,000 Publication documents
The `Publication` format has had its rendering app switched to
government-frontend. Due to the large number of documents involved we
are back filling these in batches. This is the first batch of 10,000
from most recent backwards. All newly created/published publications
will already be being rendered with government-frontend so we can work
backwards from this migration.

The data migration will output the oldest document `id` that has been
republished. This should be recorded and form the starting point for the
next migration.