Create valid url for WLNA image_details hash
When creating placeholder images, the forward slash is not automatically
added. Non placeholder images are, however,  behaving correctly.

By creating a URI joining our `public_asset_host` to our string based
`lead_image_path`, we create a valid url pointing to the image (whether
or not it is a placeholder).