Replace FindChildest with a public method
This is hard to understand as a separate class.

It takes a list of selected taxon content ids for taxons and removes any that
are redundant (those that are anscestors of a more specific selected taxon)

Add methods to EducationTaxonomyHelper to make it more obvious what the tests are doing.