Set `Consultation` rendering app to Whitehall
We need to continue render `Consultation` preview from Whitehall for now as
the format has attachments which require links.

When we implement the draft link functionality from Publishing API we
can revert this commit.

Preview relies on the `Consultation#rendering_app` so if returning
`WHITEHALL_FRONTEND` from there will have the required effect.