Prevent page crashing when edition has no expanded_links
On whitehall, when a user creates a new document, and that document is
then associated with an Education organisation (under lead
organisations) and saves the page, it crashes the next page.

This is because

>A draft can be tagged by the new taxonomy tagging form, but it doesn't
>receive other kinds of tags (set in the edit form) until publish time.
>All the tags except for taxons are coupled to the 2i workflow at the
>moment.

This commit rescues `GdsApi::HTTPNotFound ` and uses a new object,
`MissingExpandedLinks`,
this object is responsible to solely return an empty array and prevent
the view to crash.

Trello:
https://trello.com/c/nsuNcWNC/504-fix-bug-with-editing-a-draft-document-and-expanded-links