Use current government when Speech#delivered_on is future date or nil...
Speech#delvered_on is used to determine the Government attributed to the item,
if this date is nil or in the future the government won't be attributed to the
speech. This causes problems for schema validation and doesn't make sense in
the frontend. So if no delivered on date is present save the speech against
the current government, and if the delivered_on date is in the future also
assume this to be a speech from the current government.