Fix editing taxons for never published drafts
This is the same issue as af45e652b43316fcfe198dfa3d9b5e9f9034c12e
but on the tagging form itself.

Publishing API returns 404 for documents that don't have any links.  This is
always the case for brand new documents that have never been published, because
whitehall only sends the links captured at edit time when the draft is
published.