Fix placeholder News Article images presented to the Publishing API
Placeholder image URLs for News Articles and World Location News
Articles need to be generated using the correct Rails asset helpers
and configuration so that the generated URL is publicly accessible.

The host is overridden so that the frontend configuration is used.

Further work is needed so that News Articles and World Location News
Articles are republished if the lead image changes.

See Trello: https://trello.com/c/6X9zR6wW