All `Edition` formats use `PublishingApiScheduleWorker`
There was a guard clause when scheduling/unscheduling `Edition`
publishing that prevented non-migrated content from using the
`PublishingApiScheduleWorker` (which handles the reduction of cache
time) when scheduling publication. This guard checked the `Edition#rendering_app`
to ascertain the migrated status of the `Edition` format.

For the `Publication` format the `#rendering_app` has intentionally been
left returning `whitehall-frontend` as this is used to determine the
rendering app for preview and we currently can't preview this format
with government-frontend due to the lack of draft links.

This has caused an issue where scheduled publishing of `Publication`
documents appears to be delayed due to cacheing. The guard clause is no
longer required as all formats likely to be scheduled are migrated.

This commit removes it and updates the test to use `Publication` instead
of `CaseStudy` as `CaseStudy`s are not generally time sensitive and
don't make a particularly useful subject for testing this functionality.